requirements.txt
attrs==21.4.0
binance==0.3
certifi==2021.10.8
charset-normalizer==2.0.12
clearml==1.1.6
furl==2.1.3
future==0.18.2
idna==3.3
importlib-resources==5.4.0
jsonschema==4.4.0
numpy==1.22.2
orderedmultidict==1.0.1
pathlib2==2.3.7.post1
Pillow==9.0.1
psutil==5.9.0
PyJWT==2.1.0
pyparsing==3.0.7
pyrsistent==0.18.1
python-dateutil==2.8.2
PyYAML==6.0
requests==2.27.1
six==1.16.0
urllib3==1.26.8
zipp==3.7.0
python-dotenv~=0.19.2
python-binance
torch~=1.10.2
pandas~=1.4.1
stockstats~=0.4.1